# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: All-Aboard API
  description: |
    Contrat HTTP Phase 2 MVP (feed, auth JWT, demandes d'aide, mentor feed).
    Source de vérité types : `packages/types`. Doc opérationnelle : `Docs/guides/web-api-integration.md`.
  version: 0.10.0
  contact:
    name: All-Aboard THP
    url: https://github.com/AllAboard-THP/All-Aboard
servers:
  - url: http://localhost:4000
    description: Développement local
  - url: https://api-staging.allaboard.fr
    description: Staging Dokploy
...
paths:
...
  /users/me:
    patch:
      operationId: patchUsersMe
      tags: [users]
      summary: Mettre à jour le profil courant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserMeBody"
      responses:
        "200":
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserMeResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Profil public + posts ou réponses paginés
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tab
          in: query
          schema:
            type: string
            enum: [posts, responses]
            default: posts
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Profil public
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserResponse"
        "404":
          description: Utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /mentor/feed:
    get:
      operationId: getMentorFeed
      tags: [mentor]
      summary: Demandes taguées pour le mentor connecté
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Fil mentor enrichi (compteurs réponses, unread)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorFeedResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized
        "403":
          description: Rôle mentor requis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: forbidden
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mentor/dashboard:
    get:
      operationId: getMentorDashboard
      tags: [mentor]
      summary: Tableau de bord mentor (stats, file modération, aide mentor)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agrégat dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorDashboardResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Rôle mentor requis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mentor/resources/{id}/approve:
    post:
      operationId: approveMentorResource
      tags: [mentor]
      summary: Approuver une ressource en attente (compétences mentor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ressource publiée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveResourceResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Rôle mentor requis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ressource introuvable ou non modérable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mentor/resources/{id}/reject:
    post:
      operationId: rejectMentorResource
      tags: [mentor]
      summary: Refuser une ressource en attente (compétences mentor)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ressource refusée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApproveResourceResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Rôle mentor requis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ressource introuvable ou non modérable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT émis par POST /auth/login (claim sub = userId)
...
  schemas:
...
    UserRole:
      type: string
      enum: [student, mentor, admin]
    HelpRequest:
      type: object
      required: [id, title, authorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authorId:
          type: string
        authorProfileId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        body:
          type: string
        codeSnippet:
          type: string
        codeLanguage:
          type: string
        urgent:
          type: boolean
        status:
          type: string
          enum: [open, resolved]
        mentorHelpRequested:
          type: boolean
        subjectId:
          type: string
          format: uuid
        subject:
          $ref: "#/components/schemas/SubjectSummary"
        educationLevel:
          type: string
        aiSummary:
          type: string
        likesCount:
          type: integer
          minimum: 0
        responsesCount:
          type: integer
          minimum: 0
        bookmarksCount:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
        flaggedForModeration:
          type: boolean
        deletedAt:
          type: string
          format: date-time
    SubjectSummary:
      type: object
      required: [id, name, slug, icon, accentColor]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        accentColor:
          type: string
...
    Response:
      type: object
      required: [id, helpRequestId, body, authorId]
      properties:
        id:
          type: string
          format: uuid
        helpRequestId:
          type: string
          format: uuid
        body:
          type: string
        authorId:
          type: string
        authorProfileId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        codeSnippet:
          type: string
        codeLanguage:
          type: string
        flaggedForModeration:
          type: boolean
...
    MentorFeedResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MentorFeedItem"
    MentorFeedItem:
      allOf:
        - $ref: "#/components/schemas/HelpRequest"
        - type: object
          required: [responseCount, lastResponseAt, hasUnreadForMentor]
          properties:
            responseCount:
              type: integer
              minimum: 0
            lastResponseAt:
              type: string
              format: date-time
              nullable: true
            hasUnreadForMentor:
              type: boolean
    ResourceStatus:
      type: string
      enum: [pending, published, rejected]
    Resource:
      type: object
      required: [id, title, body, authorId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        authorId:
          type: string
        status:
          $ref: "#/components/schemas/ResourceStatus"
        subjectId:
          type: string
          format: uuid
        subject:
          $ref: "#/components/schemas/SubjectSummary"
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ApproveResourceResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/Resource"
...
    MentorDashboardStats:
      type: object
      required: [myResourcesCount, pendingResourcesCount, helpMentorQueueCount]
      properties:
        myResourcesCount:
          type: integer
          minimum: 0
        pendingResourcesCount:
          type: integer
          minimum: 0
        helpMentorQueueCount:
          type: integer
          minimum: 0
    MentorDashboardResponse:
      type: object
      required: [stats, myResources, pendingResources, helpMentorQueue]
      properties:
        stats:
          $ref: "#/components/schemas/MentorDashboardStats"
        myResources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        pendingResources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        helpMentorQueue:
          type: array
          items:
            $ref: "#/components/schemas/HelpRequest"
...
    UserProfile:
      type: object
      required: [id, email, role, displayName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        displayName:
          type: string
        fullName:
          type: string
        headline:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        educationLevel:
          type: string
        cguAcceptedAt:
          type: string
          format: date-time
        notifyOnComment:
          type: boolean
        notifyOnMessage:
          type: boolean
        certificationTags:
          type: array
          items:
            type: string
        competenceSubjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserMeBody:
      type: object
      properties:
        fullName:
          type: string
        headline:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        educationLevel:
          type: string
          nullable: true
        notifyOnComment:
          type: boolean
        notifyOnMessage:
          type: boolean
        subjectIds:
          type: array
          items:
            type: string
            format: uuid
        certificationTags:
          type: array
          items:
            type: string
    UpdateUserMeResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/UserProfile"
    UserPublicProfile:
      type: object
      required: [id, role, displayName, stats]
      properties:
        id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/UserRole"
        displayName:
          type: string
        headline:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        educationLevel:
          type: string
        competenceSubjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSummary"
        createdAt:
          type: string
          format: date-time
        stats:
          type: object
          required: [postsCount, responsesCount]
          properties:
            postsCount:
              type: integer
              minimum: 0
            responsesCount:
              type: integer
              minimum: 0
    PublicUserResponse:
      type: object
      required: [profile, tab, items, pagination]
      properties:
        profile:
          $ref: "#/components/schemas/UserPublicProfile"
        tab:
          type: string
          enum: [posts, responses]
        items:
          type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/HelpRequest"
              - $ref: "#/components/schemas/Response"
        pagination:
          type: object
          required: [page, limit, total]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...